+++ tests/testthat/test-verb-mutate.R
@@ -123,6 +123,28 @@ test_that("quoting for rendering mutated grouped table", {
   expect_equal(out %>% collect, tibble(x = 1, y = 1))
 })
 
+test_that("mutate generates subqueries as needed", {
+  lf <- lazy_frame(x = 1, con = simulate_sqlite())
+
+  reg <- list(
+    inplace = lf %>% mutate(x = x + 1, x = x + 1),
+    increment = lf %>% mutate(x1 = x + 1, x2 = x1 + 1)
+  )
+
+  expect_known_output(print(reg), test_path("sql/mutate-subqueries.sql"))
+})
+
+test_that("mutate collapses over nested select", {
+  lf <- lazy_frame(g = 0, x = 1, y = 2)
+
+  reg <- list(
+    xy = lf %>% select(x:y) %>% mutate(x = x * 2, y = y * 2),
+    yx = lf %>% select(y:x) %>% mutate(x = x * 2, y = y * 2)
+  )
+
+  expect_known_output(print(reg), test_path("sql/mutate-select-collapse.sql"))
+})
+
 # sql_build ---------------------------------------------------------------
 
 test_that("mutate generates simple expressions", {
@@ -132,7 +154,6 @@ test_that("mutate generates simple expressions", {
 
   expect_equal(out$select, sql(x = '`x`', y = '`x` + 1'))
 })
-
 
 # ops ---------------------------------------------------------------------